requirements.txt
requests>=2.31.0
protobuf>=4.25.1
grpcio>=1.59.3
grpcio-tools>=1.59.3
